Move shared interfaces from NPEModel to APIData
Relocated BuffersByOperation, DeviceInfo, and DeviceOperationMapping interfaces from NPEModel.ts to APIData.ts for better separation of concerns and to avoid circular imports. Updated all relevant imports in components and hooks to reference APIData.ts instead of NPEModel.ts.
