+++ src/components/OperationListPerfData.tsx
@@ -6,9 +6,8 @@ import classNames from 'classnames';
 import { Fragment } from 'react/jsx-runtime';
 import { formatSize } from '../functions/math';
 import { getCoreColour, getOpToOpGapColour } from '../functions/perfFunctions';
-import { useGetDeviceOperationListPerf } from '../hooks/useAPI';
+import { DeviceOperationMapping, useGetDeviceOperationListPerf } from '../hooks/useAPI';
 import { OperationDescription } from '../model/APIData';
-import { DeviceOperationMapping } from '../model/NPEModel';
 
 interface OperationListPerfDataProps {
     operation: OperationDescription;
