+++ src/components/buffer-summary/BufferSummaryPlotRenderer.tsx
@@ -40,7 +40,8 @@ import { L1_DEFAULT_MEMORY_SIZE } from '../../definitions/L1MemorySize';
 import { ScrollLocations } from '../../definitions/ScrollPositions';
 import useRestoreScrollPosition from '../../hooks/useRestoreScrollPosition';
 import useScrollShade from '../../hooks/useScrollShade';
-import { BuffersByOperation } from '../../model/NPEModel';
+
+import { BuffersByOperation } from '../../model/APIData';
 
 const PLACEHOLDER_ARRAY_SIZE = 50;
 const OPERATION_EL_HEIGHT = 20; // Height in px of each list item
