+++ src/components/buffer-summary/BufferSummaryTable.tsx
@@ -17,10 +17,9 @@ import useSortTable, { SortingDirection } from '../../hooks/useSortTable';
 import { TensorsByOperationByAddress } from '../../model/BufferSummary';
 import { toHex } from '../../functions/math';
 import { getBufferColor, getTensorColor } from '../../functions/colorGenerator';
-import { Buffer, BufferData } from '../../model/APIData';
+import { Buffer, BufferData, BuffersByOperation } from '../../model/APIData';
 import { selectedTensorAtom } from '../../store/app';
 import { BufferTableFilters, ColumnKeys, Columns } from '../../definitions/BufferSummary';
-import { BuffersByOperation } from '../../model/NPEModel';
 
 interface BufferSummaryTableProps {
     buffersByOperation: BuffersByOperation[];
