+++ src/model/APIData.ts
@@ -293,3 +293,46 @@ export interface BufferPage {
     tensor_id?: number;
     color?: string;
 }
+
+export interface BuffersByOperation {
+    buffers: Buffer[];
+    id: number;
+    name: string;
+}
+
+export interface DeviceInfo {
+    address_at_first_l1_bank: number;
+    address_at_first_l1_cb_buffer: number;
+    cb_limit: number;
+    device_id: number;
+    l1_bank_size: number;
+    l1_num_banks: number;
+    num_banks_per_storage_core: number;
+    num_compute_cores: number;
+    num_storage_cores: number;
+    num_x_compute_cores: number;
+    num_x_cores: number;
+    num_y_compute_cores: number;
+    num_y_cores: number;
+    total_l1_for_interleaved_buffers: number;
+    total_l1_for_sharded_buffers: number;
+    total_l1_for_tensors: number;
+    total_l1_memory: number;
+    worker_l1_size: number;
+}
+
+export interface PerformanceLog {
+    PCIe_slot: number;
+    RISC_processor_type: string; // Can we scope this down to a specific set of values?
+    core_x: number;
+    core_y: number;
+    run_ID: number;
+    run_host_ID: number;
+    source_file: string;
+    source_line: number;
+    stat_value: number;
+    'time[cycles_since_reset]': number;
+    timer_id: number;
+    zone_name: string;
+    zone_phase: 'begin' | 'end';
+}
